trimming the port
