+++ src/helpers_new/__init__.py
@@ -1,0 +1,1 @@
+"""Shared helpers for the resolve pipeline."""
