+++ src/orchestration/__init__.py
@@ -1,0 +1,1 @@
+"""Orchestration layer for resolve pipeline."""
