+++ src/sources/acled.py
@@ -1,11 +1,10 @@
-"""ACLED question source — custom resolution logic."""
+"""ACLED question source."""
 
 from __future__ import annotations
 
 import json
 import logging
 from datetime import timedelta
-from enum import Enum
 from typing import ClassVar
 
 import numpy as np
@@ -14,38 +13,18 @@ import pandas as pd
 from _schemas import AcledResolutionFrame
 from _types import SourceType
 
-from ._base import DataSource
+from ._data import DataSource
 
 logger = logging.getLogger(__name__)
 
 
-class QuestionType(Enum):
-    """Types of ACLED questions."""
-
-    N_30_DAYS_GT_30_DAY_AVG_OVER_PAST_360_DAYS = 0
-    N_30_DAYS_X_10_GT_30_DAY_AVG_OVER_PAST_360_DAYS_PLUS_1 = 1
-
-
 class AcledSource(DataSource):
-    """ACLED data source — custom row-by-row resolution logic."""
+    """Armed Conflict Location & Event Data source with custom resolution logic."""
 
     name: ClassVar[str] = "acled"
     display_name: ClassVar[str] = "ACLED"
     source_type: ClassVar[SourceType] = SourceType.DATA
-    source_intro: ClassVar[str] = (
-        "The Armed Conflict Location & Event Data Project (ACLED) collects real-time data on the "
-        "locations, dates, actors, fatalities, and types of all reported political violence and "
-        "protest events around the world. You're going to predict how questions based on this data "
-        "will resolve."
-    )
-    resolution_criteria: ClassVar[str] = (
-        "Resolves to the value calculated from the ACLED dataset once the data is published."
-    )
     resolution_schema: ClassVar[type] = AcledResolutionFrame
-
-    # ------------------------------------------------------------------
-    # Custom _resolve (port of resolve_forecasts/acled.py:resolve)
-    # ------------------------------------------------------------------
 
     def _resolve(self, df: pd.DataFrame, dfq: pd.DataFrame, dfr: pd.DataFrame) -> pd.DataFrame:
         """Resolve ACLED questions row by row."""
@@ -86,7 +65,7 @@ class AcledSource(DataSource):
         return df
 
     def _resolve_single_question(self, mid, forecast_due_date, resolution_date, dfq, dfr):
-        """Resolve an individual ACLED question."""
+        """Resolve an individual ACLED question by unhashing the ID and comparing aggregates."""
         question = self._get_question(dfq, mid)
         if question is None:
             logger.warning(f"ACLED: could NOT find {mid}")
@@ -101,13 +80,9 @@ class AcledSource(DataSource):
             resolution_date=resolution_date,
         )
 
-    # ------------------------------------------------------------------
-    # ACLED-specific resolution logic (port of helpers/acled.py)
-    # ------------------------------------------------------------------
-
     @staticmethod
     def _acled_resolve(key, dfr, country, event_type, forecast_due_date, resolution_date):
-        """Resolve given the QuestionType."""
+        """Compare 30-day sum at resolution_date against baseline at forecast_due_date."""
         lhs = AcledSource._sum_over_past_30_days(
             dfr=dfr,
             country=country,
@@ -125,7 +100,7 @@ class AcledSource(DataSource):
 
     @staticmethod
     def _sum_over_past_30_days(dfr, country, col, ref_date):
-        """Sum over the 30 days before the ref_date."""
+        """Sum of col for country over the 30 days before ref_date."""
         dfc = dfr[dfr["country"] == country].copy()
         if dfc.empty:
             return 0
@@ -137,7 +112,7 @@ class AcledSource(DataSource):
 
     @staticmethod
     def _thirty_day_avg_over_past_360_days(dfr, country, col, ref_date):
-        """Get the 30 day average over the 360 days before the ref_date."""
+        """30-day average (total/12) over the 360 days before ref_date."""
         dfc = dfr[dfr["country"] == country].copy()
         if dfc.empty:
             return 0
@@ -149,12 +124,12 @@ class AcledSource(DataSource):
 
     @staticmethod
     def _thirty_day_avg_over_past_360_days_plus_1(dfr, country, col, ref_date):
-        """Get 1 plus the 30 day average over the 360 days before the ref_date."""
+        """1 + 30-day average over the 360 days before ref_date."""
         return 1 + AcledSource._thirty_day_avg_over_past_360_days(dfr, country, col, ref_date)
 
     @staticmethod
     def _get_base_comparison_value(key, dfr, country, col, ref_date):
-        """Get the base comparison value given the question type."""
+        """Return the baseline value for comparison given the question key string."""
         if key == "last30Days.gt.30DayAvgOverPast360Days":
             return AcledSource._thirty_day_avg_over_past_360_days(
                 dfr=dfr,
@@ -184,13 +159,5 @@ class AcledSource(DataSource):
         return json.dumps(self.hash_mapping, indent=4)
 
     def _id_unhash(self, hash_key: str):
-        """Decode ACLED Ids."""
+        """Look up the original question dict from a hash key."""
         return self.hash_mapping.get(hash_key)
-
-    def fetch(self, **kwargs):
-        """Fetch raw data (stub for later phases)."""
-        raise NotImplementedError
-
-    def update(self, dfq: pd.DataFrame, **kwargs):
-        """Update questions (stub for later phases)."""
-        raise NotImplementedError
